CDP-1460
Added whitelist lib which contains a CSV of approved emails and a JS function for checking an email against that list.
Added env var WHITELISTED_EMAILS_ONLY which enables/disables the crosschecking of emails against the whitelist before login/registration.
Added isWhitelisted query and resolver for client to use to prevent having to expose the email list on the client.